Bug 791340 - Exploded WAR deployed through JON never added to inventory
When a new resource deployment takes some time, the CreateResourceRunner won't be able to discover the new resource. Eventually, the user will have to wait for next automatic service scan to see the resource in the inventory, or will have to trigger a service scan manually.

Now CreateResourceRunner will try up to five times to detect the new resource, with a 30 seconds pause between tries. That gives the managed server up to 5 minutes to deploy it.