Restore expand when context menu refreshing the selected node in the resource tree. This was broken by the fix for https://bugzilla.redhat.com/show_bug.cgi?id=816086.
Note that the behavior is still sort of odd if refreshing a node other than
the selected node.  Because we have no logic in place for actually expanding
down to a non-selected, bu right-click refreshed, node.  One possibility
would be to select the node if the user performs a right click refresh on
it, but that may be heavy handed.  For noe the refresh happens but the node
will likely be collapsed.

p.s. ignore some formatting changes that were auto-applied.