fixing a couple of (un)deployment bugs
Adding a couple missing break statement in switch blocks in the deploy/undeploy
methods in StorageNodeManagerBean. Also fixing a minor issue in
StorageNodeOperationsHandlerBean where the storage node needed to be merged
back into the persistence context so an update wasn't lost. I discovered this
when I noticed that repair wasn't getting run when it should of been during
undeployments.