adding missing exception handling for (un)deployments
There is exception handling in place for when operations complete. The
exception handle will ensure a specific error message is loged and that error
fields are set to make known that the (un)deployment failed. That exception
handling was missing in a couple placing. The first place is when the
(un)deployment is advanced in response to a cql driver event notification.
The second place is when the (un)deployment is retried.