fix subtle bug with storage client subsystem initialization
StorageClientManagerBean previously called
StorageNodeManagerBean.getStorageNodes() to get the list of hosts with which to
initializa the cql session. getStorageNodes() returns all nodes, including
those that are not yet part of the cluster. We only want clustered nodes here.

This commit also sets a round robin load balancing policy for the cql session,
and the default retry policy.