initialze storage cluster settings at install time
The storage installer now also adds the gossip port to rhq-server.properties.
The server installer now initializes the cluster settings, namely the cql and
gossip ports. This was previously done in server code when the first storage
node is imported into inventory. This simplifies things. We are able to remove
the conditional logic for initializing the cluster settings. And we no longer
need to store the cql and gossip ports in the storage node's plugin
configuration.