more fixes/improvements for storage node (un)deployment
There was a deadlock issue that could manifest itself in the
performAddNodeMaintenanceIfNecessary and in the
performRemoveNodeMaintenanceIfNecessary methods when an occurred. Both methods
had nested transactions in which both the outer and inner transactions tried to
update the same storage node entity. The transactions aren no longer nested.

This commit also adds support for continuing the (un)deployment in situations
where the event notifications for the node joining/leaving the cluster are
missed.