adding support for strongly consistent reads
This commit puts in place a strategy for the consistency level of reads and
writes such that it provides strong consistency on user-facing queries. If
the consistency level requirements cannot be satisfied, due to a down replica
for instance, the consistency level will be downgraded to provide greater
availability at the risk of possible loss of consistency.

We need to do some performance testing with this to see what tweaks need to
be made. It might also be good to make the strong consistency configurable
because some users might be fine with weaker consistency and increased
availability as well as performance on reads and writes.