fixing storage plugin itest failure
The failure occurred as a result of the upgrade to C* 1.2.9. The problem is in
test code only. In C* 1.2.9, StorageService throws an IllegalStateException if
no seed node can be contacted during bootstrap, causing the node to shut down.
The prepareForBootstrap test previously only ran with a single node and did not
really bootstrap the node.

The change in 1.2.9 forced me to make the test a bit more honest which is a
good thing. The prepareForBootstrap test now stands up a second node that is
bootstrapped.