obfuscate rhq.storage.password
In RHQ 4.8, we had rhq.cassandra.username and rhq.cassandra.password. These
been replaced with rhq.storage.username and rhq.storage.password. The
password property needs to be obfuscated like the relational database password.
If you set rhq.storage.password to an obfuscated string, the upgrade command
was replacing it with the plain-text value of rhq.cassandra.password. Since
the username/password was hard-coded as rhqadmin/rhqadmin in 4.8, we can safely
set rhq.storage.password to the obfuscated string.