[Bug 1003132 - Recovery alerts based on availability may not fire if AlertConditionCache isn't reloaded before next availability report] The issue here is the lag time between processing recovery actions (like disable of the to-be-recovered alert def, enable of the recovery alert def, etc) and the alerting cache reloads, which can be up to 30s.  The 30s [max] time is reasonable for other HA nodes, but for the server serving the agent, and\ firing the alert that triggers the recovery handling, it can be too long because the recovery action (if automated, like a restart) may happen before the cache reload, effectively missing the recovery alert.
Note that given the current architecture, which executes the alert
notifications prior to committing the alert (this should likely change),
means that the cache reload must still take place after in-band notification
processing. So, even though this fix will execute the reload as soon as
possible, it is feasible that a small window of vulnerability will still
exist if automated recovery actions are performed very quickly.

Also:
- add missing @Override annotations to related SLSBs
- add missing isDebugEnabled() protection to related SLSBs