fix oracle type mapping in native sql query
As part of the baseline calculation refactoring that took place during the C*
integration, a native sql query was introduced to fetch schedule ids. Oracle
returns BigDecimal which is why we are seeing the ClassCastExceptions in the
oracle test failures on jenkins. I updated the code to use DatabaseTypeFactory
and DatabaseType to handle the type mapping.