Bug 955363 - Uploading of file content using remote API/CLI requires many times more heap then file size
The CLI uses JBoss Remoting to communicate with the server. JBoss Remoting relies on JDK's HttpUrlConnection when configured for the HTTP transport. But HttpUrlConnection copies the full payload of the HTTP request in memory before sending it to the server. Therefore, when you send a huge file, you can easily reach the maximum heap.

Another problem was that the CLI itself was loading the file in memory as well, just to compute the file checksum.

Updated ResourceClientProxy to compute checksum on FileInputStream instead of file bytes.
Created new methods in ContentManager to upload content in fragments, updated ResourceClientProxy accordingly.