[BZ 1004050] add more error handling for advanced storage node deployments
One of the errors that can be made when deploying multiple storage nodes prior
to installing the server is failing to include all of the nodes in the
rhq.storage.nodes property in rhq-server.properties. This could have resulted
in an IndexOutOfBoundsException that was propagated up the call stack,
preventing resources (from the current inventory report) from being imported
into inventory.

Exception handling has been added along with detailed logging to deal with
this scenario.