[BZ 1006619] Writing a failing test to demonstrate the potential issue
After some discussion, it looks like the underlying problem is the issue
reported in BZ 846082. This commit reverts the changes to
Log4JLogEntryProcess.java in commit f719aa1 in order to reproduce the errors.
The test consistently fails with parsing errors in the static DateFormat
fields. If the DateFormat fields are changed back to be non-static, then the
test passes.