[BZ 1009666] Store plugin container init listeners in a list
The PC initialization listeners were introduced to work around a deadlock
issue with registering event pollers. The listeners are stored in a map in
which entries are keyed off of the class name. The only class that adds
listeners is EventContextImpl. This results in only every having one event
poller registered after PC initialization regardless of the number of event
pollers. Storing the listeners in a list resolves the issue.

(cherry picked from commit 3843704e61a48c6162dca42861fb0c44a70b3f47)