Bug 1006419 - Improve handling of obfuscated passwords for installation Use generated username/password for storage nodes.
This removes the need for the user to generate a password and manually
paste it into rhq-server.properties prior to an install.  It also generates
a username so that the username will not conflict with any rhq username, in
anticipation of a chained authentication mechanism that can use rhq auth as
a secondary login mechanism.  This will be useful for support, for example,
since the generated username/password will be internally used only.

The generated username and encoded version of the password are now also stored
as system config values in the database.  This allows the installer to use
the previously generated values for additional HA nodes.

The username/password logic used in install is:
IF set, use the values in the db system config
ELSE
   IF pre-set, use the values in rhq-server.properties (primarily for dev builds)
   ELSE generate username/password and set in rhq-server.properties

   store values in db system config

also:
- removed username/password defaults from cassandra.properties (which should
  likely go away, or minimally not be used in rhq-container.build.xml.
- added the new system config props to dbsetup and dbupgrade (2.141)
- limit generated usernames to alpha chars to avoid invalid c* identifiers
- remove logging of plain text storage password when creating superuser
- remove use of plain text storage password in temporary task file when
  creating superuser (now decodes just prior to cql statement execution)
- added ServerInstallUtil.fetchStorageClusterSettings() to fetch current
  storage settings in the db and updated persistStorageClusterSettingsIfNecessary()
  to include the new props.
- added storage property constants to ServerProperties and started using the
  constants in place of string literals in various places.