handle changes in lifecycle notifications that were introduced in upgrade to C* 1.2.9
Part of storage node (un)deployment relied on event notifications from the
driver that let the server know when a node has been added to or removed from
the cluster. Those notifcations come in the middle of the work flow and trigger
the next step in the repsective workflows. I learned that after upgrading to
C* 1.2.9, there were changes as to when the IEndpointLifecycleSubscriber
listeners get called. We are getting host added event notifications before the
node finishes bootstrapping. We do not want to start the next phase of the
deployment workflow unless and until the node has successfully bootstrapped.

This commit removes the usage of the event notifications for determining when
to start the next phase of the (un)deployment. After some code review of what
C* does during bootstrap and decommission, the prepareForBootstrap and
decommission operations in the storage plugin now include checks to make sure
the respective operations are successful.

This change ensures we do not move on to the next phase of the storage node
(un)deployment workflow prematurely. It also has the added benefit of
simplifying StorageNodeOperationsHandlerBean in multiple ways.