[BZ 996710] Only log warnings when C* writes to stderr
The installer now only logs a warning when C* writes to stderr. If there is in
fact in error, it will be caught during the post-install verification. Some
additional checks have been added to the install execution path. The cql,
gossip, and jmx ports are checked for conflicts. The installation will fail if
there is a conflict for any of those ports. Tests have been added as well.