Simplify 'rhqctl install' by removing --storage-config and --server-config
The rhq-server.properties file was already assumed to exist and be the
properties file to use during install *and* at runtime.  Allowing an override
at install-time only did not make a lot of sense and really just adds
to install confusion.  So, the option is removed and only the expected
rhq-server.properties file should be used at all points.

Analogously, we've now introduced rhq-storage.properties as an expected
properties file for storage installation.  There was already confusion
about whether this file should or should not exist, and if so, what it
should contain.  This file will now always exist, is fully commented, and
for dev builds will be preset with dev settings.

- both files are now generated in rhq-container.build.xml.
- storage_setup.groovy previously generated the dev version of the file. It is now used
  to add storage nodes in dev envs, or to generate test envs.
- add some up rhqctl up front verification that these expected files exist.
  More detailed validation of file contents is forthcoming...