Bug 1010330 - JBossAS7 server resource is not available after upgrade to jon3.2.er1
The bug was introduced with the implementation of BZ974876 "[RFE] ResourceUpgradeFacet should support upgrade of plugin config properties"

In DiscoveryBossBean, upgrade responses were filled with current resource plugin configuration if the upgrade request had nothing to update in it. But plugin config was loaded lazily and when going through the server services marshalling layer, it was turned into null. As a consequence, the server still had the right configuration, but the agent data was wrong.