[BZ 873866] - Minimizing API changes between JON 3.1.2.GA and JON 3.2.0.GA.
Commit 85f75e0 removed 2 enum fields from the server details, because they
are no longer used due to our move from DB to Cassandra for metric storage.

Nevertheless the clients might be using those fields to get back that
information. To retain the backwards compatibility, we need to keep those
enum fields, making the server details work without recompilation.
The server details will no longer contain any information for the for those
fields. The fact that some information might be missing is part of the
contract of the ServerDetails class though, so the users should be able to
handle that situation.