[BZ 873866] - Minimizing API changes between JON 3.1.2.GA and JON 3.2.0.GA.
The fix for bug 840512 changed the way we handle the value of an obfuscated
property, which rendered the obfuscate() method unused. It has been removed
which unfortunately is an API breaking change due to its protected
visibility.

While the chances of anyone using that method are miniscule we cannot break
the API. The method was re-introduced but made a NOOP so that it doesn't
break the obfuscation. The deprecation notice on the method suggests to
stop using it.