[BZ 873866] - Minimizing API changes between JON 3.1.2.GA and JON 3.2.0.GA.
Commit 26959a31d8c37bcb86dc6a845b5707a277bff3cb made the constructor of
the DatabaseQueryUtility private, which it should be, and removed an
unused class.

While the chances of anyone either instantiating the utility class or using
the absolutely useless DatabaseQueryUtility.StatementParameter are very
slim, we need to keep API back-compat.

The public constructor and the class have been reintroduced with
a deprecation notice.