numerous changes in metrics-simulator to simplify things
Metrics are now captured using the Metrics Core library which renders both
Stats.java and StatsCollector.java obsolete.

MetricsCollector has been simplified substantially. It is now seeded with a
starting schedule id and a batch size and generates batch size inserts each
time it runs.

A good bit of the simulator configuration that is specified in the json file
has been removed as well in an effort to make sure things are correct. As of
now intervals and time slices are fixed.

* raw data --> 1 minute
* 1 hour data --> 6 minutes
* 6 hour data --> 1 hour

This means that a day's worth of data is generated in one hour. Minor changes
have been made in MetricsServer to allow for configurable time slices.