fixing bugs and adding support for compressed time slices
The simulator now uses hard-coded TTLs and time slices. The duration of the
time slice for each table is as follows,

raw data --> 2.5 minutes
1 hr data --> 15 minutes
6 hr data --> 1 hour

Aggregation runs every 2.5 minutes. The execution time for aggregation can and
will exceed 2.5 minutes. I do not want the aggregator thread to block and wind
up kicking aggregation with the wrong start times. It now submits a task for
each aggregation so that the aggregator thread itself does not get delayed.