[BZ 1014524] - Fixing CriteriaQuery.iterator()
The iterator implementation assumed the page list consistent with
the page control, which might not be true, as brought to light by recent
work on PageControl.isConsistentWith() method.

The iterator implementation has been reworked to correctly iterate even
over the inconsistent results by loosening the assumption on the total
number of rows in all the pages.