Bug 1006352 - rhqctl install should only install the server and not start it unless explicitly instructed New behavior in place.  The rhqctl install and upgrade commands no longer leave the installed or upgraded services running, by default. In other words, they perform only an install or upgrade.
Each now supports a --start option which will ensure the installed or upgraded
services are running when the install or upgrade complates.

Note that for the install command this applies only to service actually
installed by that command invocation (i.e. install can not substitute for
start).

Note that for upgrade the storage service will always be left running after
upgrade if the migrate-data option is specified, to support the migration.

Note that the --agent-auto-start option has been removed from both commands
since they are superseded by the --start option.

also:
- fixed a recent regression in upgrade, now again aborts if it detects
  previously upgraded services.
- trivial, removed unused imports