[BZ 1003679] - New job to check bundle deployment completion.
The quartz job checks for the completion of a bundle deployment instead of
checking for that inline with the handling of individual resource
deployment reports. This avoids the possibility of a race condition when
the last two reports, if running simultaneously, could leave the deployment
in an IN_PROGRESS state even though all the resource deployments have
completed.