BZ 1015734 - clean up shutdown of agent
The way that thread interruption was being handled was in many cases preventing
threads from exiting.  In every case that a thread is interrupted, it is being
asked to exit, and if InterruptedException is somehow caught, the thread
interrupted state should be restored. For even Thread.sleep(), if the
thread is interrupted then the code should try to promptly exit the thread or
restore the interruption state to the caller. Otherwise, the interrupted
state is lost within the sleep.

Strictly speaking, for daemon threads, there is no reason to worry about
interrupted state, as those threads won't hang JVM shutdown, although those
threads will linger or possibly cause isuses if for example the plugin container
is restarted without restarting the JVM.

There are a few try-catch (Exception/Throwable) clauses that may catch
InterruptedException, and lose the interrupted state. I tried to correct
problematic areas.

Note that 'spurious wakeups' do not cause thread interruption. Wakeups are
really when a wait() exits prematurely, meaning no notify() was called. This is
why every wait() condition needs be checked in a loop.

I also cleaned up some of the novel ways RHQ used for notification and data
structure thread safety.  Most of this is just because the JDK has better
data structures and methods for synchronization now.

The RHQ server and Agent plugins do have similar issues but not addressed now.

Conflicts:
	modules/core/plugin-container/src/main/java/org/rhq/core/pc/inventory/AvailabilityExecutor.java