[BZ 743558] support for masking and obfuscating the system config passwords
To remain consistent with how we expose different passwords, a new ability
to set some system settings private was added and the storage-related
settings were marked as such. This is because they are not supposed to be
manipulated through the SystemManager but have different dedicated
public-facing methods.

Also, for technical reasons STORAGE_PASSWORD remained a STRING field even
though it contains a password. This is because we treat this value
specially because it can already come obfuscated from the user input and
we only deobfuscate it at use-sites, not while "in transit".

In summary, the public API methods:

* SystemManagerRemote#getSettings() - does not return any private settings,
masks password fields

* SystemManagerRemote#setSettings() - does not update any private settings,
can update password fields, obfuscating them before persisting them, so
that the DB does not contain them in clear text

3 new methods were added to SystemManagerLocal:

* getUnmaskedSystemSettings(bool includePrivateSettings)
* getObfuscatedSystemSettings(bool includePrivateSettings)
* deobfuscate(SystemSettings)

with the semantics obvious from their signature.

All the previous uses of (get|set)SystemSettings were updated to use an
appropriate local SLSB alternative so that the behavior doesn't change.
(cherry picked from commit a6f174a14bcc834827f8911adc75355a22f2a851)