Refactoring metrics index to break out each index row into multiple rows
Previously a row was created in metrics_index for each time slice of each table
for which there was data during that time slice. For example, if 1 hour data
was inserted at 15:00 for schedule id 123, then there would be a row in
metrics_index with a partition key of {six_hour_metrics:15:00}. And that row
would have a column with a value of the schedule id 123. This creates hot spots
because the entire partition for a given bucket/time slice is stored on a single
replic set. As the numnber of schedules having data for a given time slice
grows, that row in the index becomes increasingly large. Adding more nodes to
the cluster won't help.

With this commit, a time slice index (i.e., row in metrics_index) is now broken
down into multiple partitions (i.e., rows). A given time slice index will now
be distributed across nodes. The partition key is now
{bucket:schedule_offset:time_slice}. There will be at most 5,000 schedule ids
per partition. Further testing is needed to determine if that size should be
adjusted, be it smaller or larger.

This change impacts the aggregation code. Previously 3 index queries were
required, one for each of raw, 1 hour, and 6 hour data. Now multiple queries
may be required for each respective bucket. While this change requires more
reads, it is good in terms of scalability. As the overall load (in terms of)
metrics increases, this change will help achive greater througput of
aggregation particularly in a multi-node cluster.

There is one key piece missing that I will add in a subsequent commit. The max
and (optimally) min schedule ids need to be specified as arguments to the
aggregation code in order to determine the range of index partitions.