Bug 1022620 - Windows 2008 - Upgrade to JON3.2.ER3 fails when using oracle jdk7-32b The 64bit Java distribution assumes that it will run a '-server' JVM, which is required by RHQ (in part because the underlying AS7 also requires it but also because it should run as a server).  The 32bit distribution still makes a distinction between a client JVM (optimized to be lightweight) and a server JVM (optimized to perform).  Only the 32bit JDK supplies the server JVM, so for 32 Java bit we require the JDK.
By default the 32bit distribution installs under 'c:\Program Files (x86)\Java\'.
Unfortunately, our scripts failed when RHQ_JAVA_HOME was set to this
directory, because the parentheses cause issues with the script conditionals,
which (by brilliant M$ design) use parens for grouping.  So, fix #1 in this
check-in is to enable and delayed expansion of variables in the scripts, in our
logic determining where the java exe is.

On Windows we install RHQ into three windows services, one for storage, server
and agent.  By default, the (non--user) local system account runs the services.
Unfortunately, when using the 32bit distribution the java.io.tmpdir then
resolves to a locked-down temp directory under windows/system32, which
subsequently stops things from working.  The workaround is to use a user
account via RUN_AS[_ME] or to redefined java.io.tmpdir to be a valid
directory.  The latter choice is somewhat involved as you need to update
the property in all three service wrapper configurations.  This was not
as easily supported as it could be, so fix #2 is to add wrapper.inc files
for all of the components and beef up some commenting in several of our
script/conf files.

Also in this commit:
- fix the CLI scripts to use the new RHQ_JAVA_HOME and RHQ_JAVA_EXE_FILE_PATH
  properties. This had been previously missed.
- make the dev version of rhq-server-wrapper.inc apply lower mem settings.

ToDo: Wiki doco around 32bit distribution.