pass 1st and last schedule ids into aggregation
The range of schedules ids needs to be known at the time of aggregation in
order to determine all of the possible index partitions. A simple query
method has been added in MeasurementScheduleManagerBean to fetch the first and
last schedule ids. DataPurgeJob calls that method to obtain the schedule ids
that are in turn passed into MetricsServer.calculateAggregates().

This commit also fixes the logic in Aggregator for determining the number of
partittions.