[BZ 1026088] Do not query for storage node prematurely
One of the changes in the previous commit for this BZ was to change when we
fetch the node being deployment. This could cause a NoResultException in the
handleAnnounce method for the IN PROGRESS operation status. It did not fail the
deployment but caused a lot of noise in the logs that can and should otherwise
be avoided.