[1022642] Discovery queue can't display large inventory The issue is that we attempt to pull all the ADQ resources at one time.  Which overwhelms the view if there are many discovered resources.  For example, when many new agents monitoring many servers all start up and report discovery results at basically the same time.
This commit changes the approach such that we pull only the relevant platforms
first.  And then the server resources as needed, such as on tree expand.  Things
are complicated somewhat by the fact that we allow child server selection
without platform expansion and also but supporting "Select All".  Select All
will pull the child servers for each server in a separate DB request.  Whether
this is the best (or an acceptable) approach remains to be seen.

Finally, assuming we now can display a large inventory of uncommitted resources
it remains to be seen whether we can select all and import them all
successfully, as a quick look at the code reveals that we do it all
in one transaction.  It may make sense to break up the import work and
process one platform at a time.