[1022620] Windows 2008 - Upgrade to JON3.2.ER3 fails when using oracle jdk7-32b The original work tried to set override java.io.tmpdir at runtime in both the server (StartupBean) and the agent (AgentMain).  This turns out to be a bad idea as the override is not guaranteed to be used.  From the docs: " A different value may be given to this system property when the Java   virtual machine is invoked, but programmatic changes to this property   are not guaranteed to have any effect upon the the temporary directory   used..."
So, the value must be set as a -D and used in the JVM initialization.  The
default value of java.io.tmpdir is primarily an issue on Windows as we
require the 32-bit Java Service Wrapper and therefore a 32-bit JVM (unless
the user provides the 64-bit JSW on their own).

The resolution is now that for the server and agent we automatically
override java.io.tmpdir in the JSW .conf file to be %HOME%/temp.  And now
in the code, at startup, we'll create that directory if necessary.  Any
other accessibility problems with the tmpdir setting will cause an abort
at startup.