[BZ 1027458] determine whether hostname or ip address should be used
Storage nodes created by the installer can be configured to use either hostname
or ip address. When they get imported into inventory, we need make sure we
account for both. If the storage node is already in the db with a hostname but
the agent reports it with an ip address, then we want to update its address
field to use the ip address since that is how the node is actually configured.
And do the same other way around.