[BZ 1028636] Dampening broken by agent alert condition cache reloads
Reloading the cache lost the "activity" state of conditions. This caused
dampening to never receive inactivity messages and thus misbehave.

Changing the cache elements to default to "active" state solves this at the
cost of sending 1 redundant inactivity message after a cache reload. This
is harmless and has no effect on the dampening resolution.

The actual implementation changes the activity boolean to a tristate enum
of UNKNOWN, ACTIVE and INACTIVE, where UNKNOWN and ACTIVE have the same
semantics but are logically disticnt.