[BZ 1018213] Handle relative paths when purging data directories
Previously absolute paths were always used for the commitlog_directory,
data_file_directories, and saved_caches_directory properties in cassandra.yaml.
When we deploy a node, we purge those directories in order to force the node to
bootstrap from the cluster. If relative paths are used, those directories were
not getting purged; consequently, the node skipped the bootstrap step at start
up. This has the very severe consequence of resetting the system_auth keyspace.

This commit checks adds logic to handle relative paths. Paths are assumed to
be relative to <rhq-server-dir>/rhq-storage/bin.