only execute one query for unacknowledged alerts
StorageNodeManagerBean.getStorageNodeComposites() previously made separate
query for for unacked alerts for each storage node. I have added a new method
to fetch the unacked alerts in a single query. There was also a subtle bug in
the method findResourcesWithAlertsToStorageNodeMap. It is supposed to return a
map of ids of all storage node resources (or descendants) that have alert defs.
The map however included the resource ids of the entire storage node resource
hierarchy. The problem was with this code,

if (resource.getAlertDefinitions() != null) {
    resourceIdsToStorageNodeMap.put(resource.getId(), initialStorageNode.getId());
}

getAlertDefinitions() is always non-null. I have changed that check to be a call
to isEmpty().