[BZ 1027740] getStorageNodeComposites() requires MANAGE_SETTINGS permission
Adding the required permission also allows for a significant performance gain.
Prior to this commit, MeasurementDataManagerBean.getMeasurementAggregate was
invoked to fetch the load metrics. Each call incurred the additional overhead
of two queries - one to load the schedule and the other to ensure the subject
has access to view the schedule's data.

Because we now require MANAGE_SETTINGS permission, we do not need to check each
schedule. This allows us to fetch the data directly from MetricsServer and
provides a considerable speed up.

Conflicts:
	modules/enterprise/server/jar/src/main/java/org/rhq/enterprise/server/cloud/StorageNodeManagerBean.java