[BZ 1027892] clean up some exception handling and add logging
There is no need to catch NoHostAvaiableException in getLoadAsync(). Only async
queries are used, and any exceptions are propagated through the
ResultSetFuture.setException method.

I have also wrapped each of the function futures with a FutureFallback.
getLoadAsync() has a return type of
ListenableFuture<List<StorageNodeLoadComposite>>. It calls
Futures.successfulAsList() to createi a ListenableFuture that is a combination of
a list of ListenableFutures. Listeners of the combined future get called when all
of the futures in the list complete. This an extremely powerful method, but it
does not provide a hook for listeners to receive notifications when one of the
futures fails. I wrapped each future in a FutureCallback so that we can log any
error that occur when querying for data. This also ensures that the elements of
the List<StorageNodeLoadComposite> returned by the future are all non-null.