[1029553] Recovery alerts involving availability may not fire in HA environment Convert avail duration condition checking from quartz job to EJB Timer to ensure that the the job executes on the HA server initiating the job.  This server presumably still serves the relevant agent and therefore is the server which should have its global cache updated immediately upon firing of the alert, in order to quickly get recovery alert conditions into the global cache.
The only downside is that loss of the server (in an HA env) will mean loss
of the avail duration check, and potential alert.  In a non-HA env the issue
is moot because there would be no where to execute the quartz job as well.

Cherry-Pick Master: 14dca980d0b5794b5c13aebe5add5b61fd068a06