Bug 1030063 - Clean up plugin update to work synchronously
The plugin update is problematic as it is driven by listener ordering and possibly has many race conditions like seen in Bug 1025844. For example, the listener is actually added before the plugin container starts. Although effectively, the dependent systems are initialized before be the listener is called, this is not something predictable or clear.

The other weird thing is a lot of polling (of directory states, etc.) for a count of files. There is no notion of completed state, i.e. the update actually completed.

I've seen agents come up with the old version of plugins (like if the server blocked for a very long time), etc.

Using marker files isn't terribly reliable either, and not really helpful. Maybe it is a diagnostic feature, but why not use a static semaphore to order updates?

Anyway, my patch is simply having AgentMain do (more or less):

   new PluginUpdate(...).updatePlugins()

when the server starts.