[1035722] JBWEB000235: Allocate exception for servlet This is not a simple issue to resolve because it is not RHQ that logs the ERROR level exception.  It is logged by JBoss.  Our code is required to throw a ServletException indicating that we can't find the servlet server invoker (because it's not yet initialized).
We could hide ERRORs for this JBoss package via log configuration, but
then we may suppress actual, relevant runtime problems.

The commit updates the message included in our ServletException to be
very explicit about this not being an actual problem during server
startup and initialization.  Hopefully this will be sufficient.

If this proves insufficient we should likely suppress the logging or, if
we absolutely must log it, direct it to another file (not really desirable).